+++ platforms/react/src/utils/getMediaQueries/index.ts
@@ -1,6 +1,6 @@
 import type { DefaultTheme } from 'styled-components'
 import type { ThemeGetterValue } from '@shallot-ui/core-utils'
-import type { ThemeOptions } from '@shallot-ui/core-theme'
+import type { ThemeTokens } from '@shallot-ui/core-theme'
 
 import { makeTheme } from '@shallot-ui/core-theme'
 
@@ -11,7 +11,7 @@ export const getMediaQueries = (
   theme.breakpoints
     ? Object.entries(theme.breakpoints)
         .map(([width, subtheme]) => {
-          const style = getStyle(makeTheme(subtheme as ThemeOptions))
+          const style = getStyle(makeTheme(subtheme as ThemeTokens))
           if (style) return `@media (min-width: ${width}px) { ${style} }`
         })
         .filter(Boolean)