+++ components/Switch/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @shallot-ui/component-switch
 
+## 3.0.0-beta.2
+
+### Minor Changes
+
+- Renames theme options to tokens for clarity
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-mixins@2.0.0-beta.2
+  - @shallot-ui/core-theme@2.0.0-beta.2
+  - @shallot-ui/core-utils@2.0.0-beta.2
+
 ## 3.0.0-beta.1
 
 ### Minor Changes