+++ core/mixins/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @shallot-ui/core-mixins
 
+## 2.0.0-beta.2
+
+### Minor Changes
+
+- Renames theme options to tokens for clarity
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/core-theme@2.0.0-beta.2
+
 ## 2.0.0-beta.1
 
 ### Minor Changes