+++ packages/native/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @shallot-ui/native
 
+## 3.0.0-beta.2
+
+### Minor Changes
+
+- Renames theme options to tokens for clarity
+
+### Patch Changes
+
+- Updated dependencies
+  - @shallot-ui/platform-react-native@2.0.0-beta.2
+  - @shallot-ui/component-checkbox@3.0.0-beta.2
+  - @shallot-ui/component-button@3.0.0-beta.2
+  - @shallot-ui/component-switch@3.0.0-beta.2
+  - @shallot-ui/component-input@3.0.0-beta.2
+  - @shallot-ui/component-tag@3.0.0-beta.2
+  - @shallot-ui/core-mixins@2.0.0-beta.2
+  - @shallot-ui/core-props@2.0.0-beta.2
+  - @shallot-ui/core-theme@2.0.0-beta.2
+  - @shallot-ui/core-utils@2.0.0-beta.2
+
 ## 3.0.0-beta.1
 
 ### Minor Changes